requirements.txt
serpy==0.3.1
colander==1.7.0
djangorestframework==3.14.0
django==4.2.7
pydantic==2.5.2
pytest
pystrainer==1.3.0
tabulate==0.9.0
lollipop==1.1.7
py-kim==1.2.2
lima==0.5
markdown
avro-python3==1.8.2
serpyco==1.1.0
marshmallow==3.20.1
